Rename to "vfs".
