+++ ailets-rs/actor_runtime_mocked/src/lib.rs
@@ -1,5 +1,3 @@
-mod actor_runtime_mocked;
+pub mod vfs;
 
-pub use actor_runtime_mocked::{
-    clear_mocks, add_file, get_file, IO_INTERRUPT, WANT_ERROR,
-};
+pub use vfs::{add_file, clear_mocks, get_file, IO_INTERRUPT, WANT_ERROR};
